requirements.txt
discord.py
pytz
Pillow
aiohttp
requests
dotenv
pydantic-ai-slim[openrouter]
pytest
pytest-asyncio
